#1344 Checking if the branch exists before deleting it

+++ ontrack-extension-github/src/main/java/net/nemerosa/ontrack/extension/github/scm/GitHubSCMExtension.kt
@@ -111,7 +111,9 @@ class GitHubSCMExtension(
         }
 
         override fun deleteBranch(branch: String) {
-            client.deleteBranch(repository, branch)
+            if (client.getBranchLastCommit(repository, branch, retryOnNotFound = false) != null) {
+                client.deleteBranch(repository, branch)
+            }
         }
 
         override fun createBranch(sourceBranch: String, newBranch: String): String {
